Refactor plugin manager into focused runtime modules
