+++ src/genecoder/plugin_runtime/__init__.py
@@ -1,0 +1,1 @@
+"""Internal plugin runtime components."""
